+++ src/version3/jqlFunctionsApps.ts
@@ -1,7 +1,7 @@
 import * as Models from './models';
 import * as Parameters from './parameters';
-import { Callback } from '../callback';
 import { Client } from '../clients';
+import { Callback } from '../callback';
 import { RequestConfig } from '../requestConfig';
 
 export class JqlFunctionsApps {
@@ -13,6 +13,9 @@ export class JqlFunctionsApps {
    *
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v3/intro/#permissions) required:** This
    * API is only accessible to apps and apps can only inspect their own functions.
+   *
+   * The new `read:app-data:jira` OAuth scope is 100% optional now, and not using it won't break your app. However, we
+   * recommend adding it to your app's scope list because we will eventually make it mandatory.
    */
   async getPrecomputations<T = Models.PageJqlFunctionPrecomputation>(
     parameters: Parameters.GetPrecomputations | undefined,
@@ -24,6 +27,9 @@ export class JqlFunctionsApps {
    *
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v3/intro/#permissions) required:** This
    * API is only accessible to apps and apps can only inspect their own functions.
+   *
+   * The new `read:app-data:jira` OAuth scope is 100% optional now, and not using it won't break your app. However, we
+   * recommend adding it to your app's scope list because we will eventually make it mandatory.
    */
   async getPrecomputations<T = Models.PageJqlFunctionPrecomputation>(
     parameters?: Parameters.GetPrecomputations,
@@ -41,7 +47,6 @@ export class JqlFunctionsApps {
         startAt: parameters?.startAt,
         maxResults: parameters?.maxResults,
         orderBy: parameters?.orderBy,
-        filter: parameters?.filter,
       },
     };
 
@@ -53,6 +58,9 @@ export class JqlFunctionsApps {
    *
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v3/intro/#permissions) required:** An API
    * for apps to update their own precomputations.
+   *
+   * The new `write:app-data:jira` OAuth scope is 100% optional now, and not using it won't break your app. However, we
+   * recommend adding it to your app's scope list because we will eventually make it mandatory.
    */
   async updatePrecomputations<T = void>(
     parameters: Parameters.UpdatePrecomputations | undefined,
@@ -63,6 +71,9 @@ export class JqlFunctionsApps {
    *
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v3/intro/#permissions) required:** An API
    * for apps to update their own precomputations.
+   *
+   * The new `write:app-data:jira` OAuth scope is 100% optional now, and not using it won't break your app. However, we
+   * recommend adding it to your app's scope list because we will eventually make it mandatory.
    */
   async updatePrecomputations<T = void>(parameters?: Parameters.UpdatePrecomputations, callback?: never): Promise<T>;
   async updatePrecomputations<T = void>(
@@ -72,8 +83,57 @@ export class JqlFunctionsApps {
     const config: RequestConfig = {
       url: '/rest/api/3/jql/function/computation',
       method: 'POST',
+      params: {
+        skipNotFoundPrecomputations: parameters?.skipNotFoundPrecomputations,
+      },
       data: {
         values: parameters?.values,
+      },
+    };
+
+    return this.client.sendRequest(config, callback);
+  }
+
+  /**
+   * Returns function precomputations by IDs, along with information about when they were created, updated, and last
+   * used. Each precomputation has a `value` - the JQL fragment to replace the custom function clause with.
+   *
+   * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v3/intro/#permissions) required:** This
+   * API is only accessible to apps and apps can only inspect their own functions.
+   *
+   * The new `read:app-data:jira` OAuth scope is 100% optional now, and not using it won't break your app. However, we
+   * recommend adding it to your app's scope list because we will eventually make it mandatory.
+   */
+  async getPrecomputationsByID<T = Models.JqlFunctionPrecomputationGetByIdResponse>(
+    parameters: Parameters.GetPrecomputationsByID | undefined,
+    callback: Callback<T>,
+  ): Promise<void>;
+  /**
+   * Returns function precomputations by IDs, along with information about when they were created, updated, and last
+   * used. Each precomputation has a `value` - the JQL fragment to replace the custom function clause with.
+   *
+   * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v3/intro/#permissions) required:** This
+   * API is only accessible to apps and apps can only inspect their own functions.
+   *
+   * The new `read:app-data:jira` OAuth scope is 100% optional now, and not using it won't break your app. However, we
+   * recommend adding it to your app's scope list because we will eventually make it mandatory.
+   */
+  async getPrecomputationsByID<T = Models.JqlFunctionPrecomputationGetByIdResponse>(
+    parameters?: Parameters.GetPrecomputationsByID,
+    callback?: never,
+  ): Promise<T>;
+  async getPrecomputationsByID<T = Models.JqlFunctionPrecomputationGetByIdResponse>(
+    parameters?: Parameters.GetPrecomputationsByID,
+    callback?: Callback<T>,
+  ): Promise<void | T> {
+    const config: RequestConfig = {
+      url: '/rest/api/3/jql/function/computation/search',
+      method: 'POST',
+      params: {
+        orderBy: parameters?.orderBy,
+      },
+      data: {
+        precomputationIDs: parameters?.precomputationIDs,
       },
     };