+++ src/version3/parameters/bulkFetchIssues.ts
@@ -1,0 +1,3 @@
+import { BulkFetchIssueRequest } from '../models';
+
+export interface BulkFetchIssues extends BulkFetchIssueRequest {}