+++ src/version3/parameters/getBulkChangelogs.ts
@@ -1,0 +1,3 @@
+import { BulkChangelogRequest } from '../models';
+
+export interface GetBulkChangelogs extends BulkChangelogRequest {}