+++ src/version3/parameters/searchAndReconsileIssuesUsingJqlPost.ts
@@ -1,0 +1,3 @@
+import { SearchAndReconcileRequest } from '../models';
+
+export interface SearchAndReconsileIssuesUsingJqlPost extends SearchAndReconcileRequest {}