+++ src/version3/parameters/searchForIssuesIds.ts
@@ -1,0 +1,3 @@
+import { IdSearchRequestBean } from '../models';
+
+export interface SearchForIssuesIds extends IdSearchRequestBean {}