+++ src/version3/parameters/validateUpdateWorkflows.ts
@@ -1,0 +1,3 @@
+import { WorkflowUpdateValidateRequestBean } from '../models';
+
+export interface ValidateUpdateWorkflows extends WorkflowUpdateValidateRequestBean {}